added files from old joana key refactoring project and implemented Window to show Java Code (using an external project)
This will now be the main project where even the old refactoring will
take place. The previous project will not be updated any more. The 
refactored joana key project will be used by the gui to let the user
interact with the prove process.

The code are uses a demo of a lib called richtext.