started working on interface between gui and disproving tools.
However, it turned out that the sdg serializer apparently
doesnt save the sdgs ability to get corres call grap ids, which
is a problem. If I cant solve it i will just have to generate the sdg
from scratch every time. Not too horrible, considering that the real bottle-
neck is always generating the violationswrapper.

It still sucks though